Revert guest RAM to 256 MiB; keep Start catch log
Julian: 2 GiB / 2047 MiB alloc still blocked Start. Restore the
original 256 MiB RamDevice and SP 0x8FFFF000. Drop the kseg0Ram
workaround and very-large-objects plumbing. Keep Start:
<type>: <message> so the next failure is not silent. Leave/ROMHDR
dump-true from 07f8a9a stays.

+++ RamDevice.cs
@@ -12,7 +12,7 @@ namespace ProcessorEmulator.Emulation
         {
             StartAddress = startAddress;
             Size = size;
-            _memory = new byte[(long)size];
+            _memory = new byte[size];
         }
 
         public uint Read32(uint offset)
